+++ src/item.cpp
@@ -12,23 +12,13 @@ Item::~Item() {
 
 }
 
-// GETTER
-bool Item::isUsed() {
-    return used;
-}
-
-// SETTER
-void Item::setUsed(bool value) {
-    used = value;
-}
-
 
 
 // BLOCK
 
 // INIT
 Block::Block(Game* game, BufferedBlock* d) : Item(game), data(d) {
-    
+
 }
 
 // CLEANUP
@@ -38,24 +28,31 @@ Block::~Block() {
 
 // UPDATE
 void Block::update() {
-
+    H2DE_TickTimelineManager(tm);
 }
 
 // RENDER
 void Block::render() {
-    renderTexture();
-    if (game->getMegahack()->getHack("show-hitboxes")->active) renderHitbox();
+    LevelPos camPos = game->getCamera()->getPos();
+
+    if (data->pos.x + data->textureSize.w >= camPos.x && data->pos.x - data->textureSize.w <= camPos.x + BLOCKS_ON_WIDTH + 1) {
+        renderTexture();
+        if (game->getMegahack()->getHack("show-hitboxes")->active) renderHitbox();
+    }
 }
 
 void Block::renderTexture() {
     static H2DE_Engine* engine = game->getEngine();
+    static GameData* gameData = game->getData();
     static Calculator* calculator = game->getCalculator();
+    Level* level = game->getLevel();
+    LevelPos playerPos = level->getPlayer()->getPos();
     LevelPos camPos = game->getCamera()->getPos();
 
     bool textureOnCamX = (data->pos.x + data->textureOffset.x + data->textureSize.w > camPos.x) && (data->pos.x + data->textureOffset.x < camPos.x + BLOCKS_ON_WIDTH + 1);
-
     if (textureOnCamX) {
         H2DE_Size absBlockSize = calculator->convertToPx(LevelSize{ 1.0f, 1.0f });
+
         H2DE_GraphicObject* texture = new H2DE_GraphicObject();
         texture->type = IMAGE;
         texture->texture = data->texture;
@@ -64,7 +61,7 @@ void Block::renderTexture() {
         texture->rotation = data->rotation;
         texture->rotationOrigin = { absBlockSize.w / 2, absBlockSize.h / 2 };
         texture->flip = data->flip;
-        texture->color = static_cast<H2DE_Color>(game->getLevel()->getData()->colors[data->colorID]);
+        texture->color = static_cast<H2DE_Color>(level->getData()->colors[data->colorID]);
         texture->index = data->zIndex->getIndex();
         H2DE_AddGraphicObject(engine, texture);
     }
@@ -77,7 +74,6 @@ void Block::renderHitbox() {
     LevelPos camPos = game->getCamera()->getPos();
 
     bool hitboxOnCamX = (data->pos.x + data->hitboxOffset.x + data->hitboxSize.w > camPos.x) && (data->pos.x + data->hitboxOffset.x < camPos.x + BLOCKS_ON_WIDTH + 1);
-    
     if (hitboxOnCamX && data->type != DECORATION) {
         H2DE_Size absHitboxSize = calculator->convertToPx(data->hitboxSize);
         H2DE_Pos absHitboxOffset = calculator->convertToPx(data->hitboxOffset);
@@ -98,7 +94,36 @@ void Block::renderHitbox() {
 
 // RESET
 void Block::reset() {
-    setUsed(false);
+    H2DE_ClearTimelineManager(tm);
+    used = false;
+}
+
+// ACTIONS
+void Block::enter() {
+    Player* player = game->getLevel()->getPlayer();
+    used = true;
+
+    switch (data->specialData.type) {
+        case SD_PORTAL: switch (data->specialData.desc) {
+            case SD_CUBE: player->setGamemode(CUBE, data->pos.y); break;
+            case SD_SHIP: player->setGamemode(SHIP, data->pos.y); break;
+            case SD_RIGHT_SIDE_UP: break;
+            case SD_UPSIDE_DOWN: break;
+        } break;
+        case SD_ORB: switch (data->specialData.desc) {
+            case SD_PINK: break;
+            case SD_YELLOW: break;
+            case SD_BLUE: break;
+        } break;
+        case SD_PAD: switch (data->specialData.desc) {
+            case SD_PINK: break;
+            case SD_YELLOW: break;
+            case SD_BLUE: break;
+        } break;
+        case SD_COIN: switch (data->specialData.desc) {
+            case SD_SECRET: break;
+        } break;
+    }
 }
 
 // GETTER
@@ -106,13 +131,17 @@ BufferedBlock* Block::getData() {
     return data;
 }
  
+bool Block::entered() {
+    return used;
+}
+
 
 
 // TRIGGER
 
 // INIT
 Trigger::Trigger(Game* game, BufferedTrigger* d) : Item(game), data(d) {
-    if (data->type == STARTPOS) setUsed(true);
+    if (data->type == STARTPOS) used = true;
 }
 
 // CLEANUP
@@ -126,7 +155,7 @@ void Trigger::update() {
 
     H2DE_TickTimelineManager(tm);
 
-    if (isUsed()) return;
+    if (used) return;
     if (player->getPos().x >= data->pos.x) {
         if (data->touchTrigger) {
 
@@ -138,7 +167,14 @@ void Trigger::trigger() {
     static H2DE_Engine* engine = game->getEngine();
     Level* level = game->getLevel();
 
-    setUsed(true);
+    for (Item* item : *(level->getItems())) {
+        Trigger* trigger = ItemManager::castToTrigger(item);
+        if (!trigger) continue;
+        if (trigger->data->pos.x > data->pos.x) break;
+        if (trigger->data->type == data->type) H2DE_ClearTimelineManager(trigger->tm);
+    }
+
+    used = true;
 
     Color sCol;
     switch (data->type) {
@@ -175,7 +211,7 @@ void Trigger::render() {
 
 // RESET
 void Trigger::reset() {
-    if (data->type != STARTPOS) setUsed(false);
+    if (data->type != STARTPOS) used = false;
     H2DE_ClearTimelineManager(tm);
 }